@@ -12,24 +12,18 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@mui/material": "^5.15.14",
-    "@prisma/client": "^5.11.0",
-    "@trpc/client": "^10.45.2",
-    "@trpc/next": "^10.45.2",
-    "@trpc/react": "^9.27.4",
-    "@trpc/server": "^10.45.2",
+    "@tanstack/react-query": "^5.29.2",
     "cuid": "^3.0.0",
     "date-fns": "^3.6.0",
     "express": "^4.19.2",
     "framer-motion": "^11.0.25",
     "lucide-react": "^0.363.0",
+    "mongodb": "^6.5.0",
+    "mongoose": "^8.3.1",
     "next": "14.1.4",
     "next-auth": "^4.24.7",
-    "nodemailer": "^6.9.13",
     "postmark": "^4.0.2",
     "react": "^18",
-    "react-beautiful-dnd": "^13.1.1",
-    "react-dnd": "^16.0.1",
-    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.22.3",
@@ -38,15 +32,15 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/node": "^20",
+    "@types/node": "^20.12.7",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
     "eslint-config-next": "14.1.4",
     "postcss": "^8",
-    "prisma": "^5.11.0",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   }
 }
